pyproject.toml
[project]
name = "chatbot"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beaver-db>=1.3.0",
    "fastembed>=0.7.4",
    "lingo-ai>=0.3.1",
    "purely>=0.1.2",
    "pyarrow>=22.0.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chatbot"]


[project.scripts]
bot = "chatbot.app:start_bot"
index = "chatbot.index:index"
